requirements.txt
flask>=3.0.0
reportlab>=4.0.0
python-docx>=1.1.0
mammoth>=1.6.0
gunicorn>=21.0.0
